+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 
 1. [#155](https://github.com/InfluxCommunity/influxdb3-csharp/pull/155): Allows setting grpc options.
+1. [#157](https://github.com/InfluxCommunity/influxdb3-csharp/pull/157): Fix: always clone `DefaultOptions` to keep it
+   immutable.
 
 ## 1.1.0 [2025-03-26]